refactor(rsbinder-tools): rename the policy module and flag to config
The directory rsb_hub reads is about to hold more than access rules: a
service declaration says who may use a service, that it exists, and how to
start it, and splitting those across two directories would mean editing two
places per service. Once `[[service]]` lands beside `[[rule]]`, calling the
whole thing "policy" fights the file's contents.

Rename now, while nothing is released and the diff is mechanical, rather
than leaving a misnomer in a security-relevant CLI where it would never get
fixed. `--policy` becomes `--config`, `rsbinder_tools::policy` becomes
`rsbinder_tools::config`, and the access-rule types move to
`config::policy` — still policy, now a part rather than the whole.

No behaviour change.
